% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burndown.R
\name{aufwand}
\alias{aufwand}
\title{Calculates the effort of a given card (given by its description)}
\usage{
aufwand(description)
}
\arguments{
\item{description}{A string extracted from an UpWave card}
}
\value{
An integer
}
\description{
aufwand() parses the description of a card for tags like "Aufwand: 10 AT" or
"Aufwand: 10 SP".
}
