% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board2excel.R
\name{board2excel}
\alias{board2excel}
\title{Save all information from a board to Excel}
\usage{
board2excel(organisation, board_id, token, filename, path)
}
\arguments{
\item{organisation}{A character}

\item{board_id}{A character}

\item{token}{A character}

\item{filename}{A character}

\item{path}{A character}
}
\description{
Save all information from a board to Excel
}
\examples{
\dontrun{
organisation <- "4597"
board_id <- 14351
token_sqc <- "807a0c3451c04602b4dbfdc0338a65f7"
upwaver::board2excel(organisation, board_id, token_sqc, "test.xlsx", "C:/Temp/")
}
}
