% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upwaver.R
\name{content_as_list}
\alias{content_as_list}
\title{Transform the answer of a 'Upwave' API request to a list}
\usage{
content_as_list(response)
}
\arguments{
\item{response}{An object of class "response"; response of a GET request}
}
\value{
A list
}
\description{
content_as_list() is an internal function that extracts content from a request and
coerces it to a list format
}
