% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upwaver.R
\name{list_boards}
\alias{list_boards}
\title{Information about your organisation's boards}
\usage{
list_boards(organisation, token)
}
\arguments{
\item{organisation}{A character. The name of the organisation as specified in the URL:
https://<organisation>.upwave.io}

\item{token}{A character. The API token to access boards on upwave.io. The token can be
obtained by visiting your account settings. Click your profile image, select "Settings"
and find your API-Key in the "Account" tab.}
}
\value{
A list
}
\description{
list_boards() is a wrapper around "List Boards" as specified in the 'Upwave' API
https://www.upwave.io/api/. It returns information for your organisation's
boards as id's, title's, time of creation and the users who created the boards.
}
\examples{
list_boards("4597", "807a0c3451c04602b4dbfdc0338a65f7")
}
