% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upwaver.R
\name{request}
\alias{request}
\title{Create a correct request for the 'Upwave' API}
\usage{
request(url, token)
}
\arguments{
\item{url}{an URL}

\item{token}{an authentication token}
}
\value{
request
}
\description{
request() is an internal function that invokes a GET request with a token according to
https://www.upwave.io/api/
}
